loading_794DC.c up to LOAD_BIG_RAYMAN

+++ include/loading_794DC.h
@@ -3,6 +3,7 @@
 
 #include "common.h"
 #include "psyq_3_0/LIBCD.H"
+#include "psyq_3_0/LIBGPU.H"
 
 typedef struct FileInfo
 {
@@ -39,10 +40,10 @@ extern s32 PS1_FileTemp;
 void PS1_Init_ImgLdrVdoTrk_Files(void);
 void FUN_8019dd74(void);
 void FUN_8019deac(void);
-void FUN_8019df1c(int param_1);
+void FUN_8019df1c(s16 param_1);
 void LOAD_BIG_RAYMAN(void);
 void LOAD_ALL_FIX(void);
-void load_world(short worldIndex);
+void load_world(s16 worldIndex);
 void load_level(void);
 void LOAD_FND(void);
 void PS1_LoadImgWorld(void);
